Update jit_generate to not use gen_patch_increment_instruction
Using the patch function allowed us to keep eip updates before function calls,
for all kinds of instructions, by using the saved offset.

With the instruction_body buffer, we don't _really_ need to be saving the
offset, since we can order eip updates and instruction body updates as we
please.

In prep for getting rid of offset_increment_instruction_pointer, this commit
lets us update eip only when we need to:

- When previous_ip needs to be set (could be optimized further; this commit
  updates eip first, and then sets previous_ip to that value, instead of just
  updating previous_ip directly)
- When a jump instruction may need to use eip