codegen/gen_safe_write32: Use local-variable based arg passing
Stack based passing was needless given that we wanted the values in
local variables anyway and could simply have the caller do it.

We could have a minor optimization and have the caller gen_tee_local for
the address, but I chose not to do that because it seems like a minor
difference that complicates the calling convention
significantly (i.e. both set local variable and push to stack).