Add tests for virtio 9p (tests are currently incorrect)
These tests do not work yet: it fails to capture the serial output
after root login.

In addition, these tests are currently too slow to run
since it takes a while to boot the debian image (> 16 minutes on my
machine while running the test in the terminal.)

Would need to switch to a newer minimal linux image when available.